backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
requests==2.31.0
APScheduler==3.10.4
python-json-logger==2.0.7

